Refactor the renderer classes and interface.
