+++ flow-components-parent/demo-flow-components/src/main/java/com/vaadin/flow/demo/views/GridView.java
@@ -42,7 +42,7 @@ import com.vaadin.ui.html.Div;
 import com.vaadin.ui.html.Label;
 import com.vaadin.ui.layout.HorizontalLayout;
 import com.vaadin.ui.layout.VerticalLayout;
-import com.vaadin.ui.renderers.ComponentRenderer;
+import com.vaadin.ui.renderers.ComponentTemplateRenderer;
 import com.vaadin.ui.renderers.TemplateRenderer;
 import com.vaadin.ui.textfield.TextField;
 
@@ -182,7 +182,7 @@ public class GridView extends DemoView {
 
         /**
          * Sets the person for the component.
-         * 
+         *
          * @param person
          *            the person to be inside inside the cell
          */
@@ -199,7 +199,7 @@ public class GridView extends DemoView {
 
         /**
          * Constructor that takes a Person as parameter.
-         * 
+         *
          * @param person
          *            the person to be used inside the card
          */
@@ -239,7 +239,7 @@ public class GridView extends DemoView {
         createColumnTemplate();
         createDetailsRow();
         createColumnGroup();
-        createColumnComponentRenderer();
+        createColumnComponentTemplateRenderer();
         createSorting();
         createHeaderAndFooterUsingTemplates();
         createHeaderAndFooterUsingComponents();
@@ -425,19 +425,19 @@ public class GridView extends DemoView {
                 grid);
     }
 
-    private void createColumnComponentRenderer() {
+    private void createColumnComponentTemplateRenderer() {
         // begin-source-example
         // source-example-heading: Grid with columns using component renderer
         Grid<Person> grid = new Grid<>();
         grid.setItems(createItems());
 
         // You can use a constructor and a separate setter for the renderer
-        grid.addColumn(new ComponentRenderer<>(PersonComponent::new,
+        grid.addColumn(new ComponentTemplateRenderer<>(PersonComponent::new,
                 PersonComponent::setPerson)).setHeader("Person");
 
         // Or you can use an ordinary function to get the component
-        grid.addColumn(
-                new ComponentRenderer<>(item -> new Button("Remove", evt -> {
+        grid.addColumn(new ComponentTemplateRenderer<>(
+                item -> new Button("Remove", evt -> {
                     ListDataProvider<Person> dataProvider = (ListDataProvider<Person>) grid
                             .getDataProvider();
                     dataProvider.getItems().remove(item);
@@ -445,7 +445,8 @@ public class GridView extends DemoView {
                 }))).setHeader("Actions");
 
         // Item details can also use components
-        grid.setItemDetailsRenderer(new ComponentRenderer<>(PersonCard::new));
+        grid.setItemDetailsRenderer(
+                new ComponentTemplateRenderer<>(PersonCard::new));
 
         // When items are updated, new components are generated
         TextField idField = new TextField("", "Person id");
@@ -675,7 +676,7 @@ public class GridView extends DemoView {
         grid.setItems(getItems());
 
         Column<Person> nameColumn = grid.addColumn(Person::getName)
-                .setHeader(new ComponentRenderer<>(() -> {
+                .setHeader(new ComponentTemplateRenderer<>(() -> {
                     Label label = new Label("Name");
                     label.getStyle().set("color", "green");
                     label.setTitle("Name");
@@ -683,7 +684,7 @@ public class GridView extends DemoView {
                 })).setComparator((p1, p2) -> p1.getName()
                         .compareToIgnoreCase(p2.getName()));
         Column<Person> ageColumn = grid.addColumn(Person::getAge, "age")
-                .setHeader(new ComponentRenderer<>(() -> {
+                .setHeader(new ComponentTemplateRenderer<>(() -> {
                     Label label = new Label("Age");
                     label.getStyle().set("color", "blue");
                     label.setTitle("Age");
@@ -698,12 +699,12 @@ public class GridView extends DemoView {
 
         ColumnGroup informationColumnGroup = grid
                 .mergeColumns(nameColumn, ageColumn)
-                .setHeader(new ComponentRenderer<>(() -> {
+                .setHeader(new ComponentTemplateRenderer<>(() -> {
                     Label label = new Label("Basic Information");
                     label.getStyle().set("color", "orange");
                     label.setTitle("Basic Information");
                     return label;
-                })).setFooter(new ComponentRenderer<>(() -> {
+                })).setFooter(new ComponentTemplateRenderer<>(() -> {
                     Label label = new Label(
                             "Total: " + getItems().size() + " people");
                     label.getStyle().set("color", "red");
@@ -711,7 +712,7 @@ public class GridView extends DemoView {
                 }));
         ColumnGroup addressColumnGroup = grid
                 .mergeColumns(streetColumn, postalCodeColumn)
-                .setHeader(new ComponentRenderer<>(() -> {
+                .setHeader(new ComponentTemplateRenderer<>(() -> {
                     Label label = new Label("Address Information");
                     label.setTitle("Address Information");
                     return label;
