+++ flow-components-parent/flow-components/src/main/java/com/vaadin/ui/grid/GridTemplateRendererUtil.java
@@ -29,7 +29,7 @@ import com.vaadin.function.SerializableConsumer;
 import com.vaadin.ui.Component;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.grid.Grid.GridDataGenerator;
-import com.vaadin.ui.renderers.ComponentRenderer;
+import com.vaadin.ui.renderers.ComponentTemplateRenderer;
 import com.vaadin.ui.renderers.ComponentRendererUtil;
 import com.vaadin.ui.renderers.TemplateRenderer;
 import com.vaadin.util.JsonSerializer;
@@ -45,7 +45,7 @@ import com.vaadin.util.JsonSerializer;
 class GridTemplateRendererUtil {
 
     /**
-     * Internal object to hold {@link ComponentRenderer}s and their generated
+     * Internal object to hold {@link ComponentTemplateRenderer}s and their generated
      * {@link Component}s together.
      * 
      * @param <T>
@@ -53,7 +53,7 @@ class GridTemplateRendererUtil {
      */
     static final class RendereredComponent<T> implements Serializable {
         private Component component;
-        private ComponentRenderer<? extends Component, T> componentRenderer;
+        private ComponentTemplateRenderer<? extends Component, T> componentRenderer;
 
         /**
          * Default constructor.
@@ -64,7 +64,7 @@ class GridTemplateRendererUtil {
          *            the renderer that generated the component
          */
         public RendereredComponent(Component component,
-                ComponentRenderer<? extends Component, T> componentRenderer) {
+                ComponentTemplateRenderer<? extends Component, T> componentRenderer) {
             this.component = component;
             this.componentRenderer = componentRenderer;
         }
@@ -80,7 +80,7 @@ class GridTemplateRendererUtil {
 
         /**
          * Recreates the component by calling
-         * {@link ComponentRenderer#createComponent(Object)}, and sets the
+         * {@link ComponentTemplateRenderer#createComponent(Object)}, and sets the
          * internal component returned by {@link #getComponent()}.
          * 
          * @param item
@@ -179,7 +179,7 @@ class GridTemplateRendererUtil {
     }
 
     static <T> void setupHeaderOrFooterComponentRenderer(Component owner,
-            ComponentRenderer<? extends Component, T> componentRenderer) {
+            ComponentTemplateRenderer<? extends Component, T> componentRenderer) {
         Element container = ComponentRendererUtil
                 .createContainerForRenderers(owner);
 
@@ -195,7 +195,7 @@ class GridTemplateRendererUtil {
     }
 
     static <T> void registerRenderedComponent(
-            ComponentRenderer<? extends Component, T> componentRenderer,
+            ComponentTemplateRenderer<? extends Component, T> componentRenderer,
             Map<String, RendereredComponent<T>> renderedComponents,
             Element container, String key, Component component) {
         component.getElement().setAttribute("data-flow-renderer-item-key", key);
